Rename : profileCard 폴더 components로 이동
