+++ package.json
@@ -23,7 +23,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/themes": "^2.0.3",
-    "@supabase/supabase-js": "^2.42.5",
+    "@supabase/supabase-js": "^2.43.0",
     "@tanstack/react-query": "^5.28.6",
     "@tanstack/react-query-devtools": "^5.28.6",
     "@tanstack/react-table": "^8.15.0",
